refactor(cli): extract run get subcommands to run_get.go (PRD #1009 M4)
